n-api: optimize number API performance
 - Add separate APIs for creating different kinds of numbers,
   because creating a V8 number value from an integer is faster
   than creating one from a double.
 - When getting number values, avoid getting the current context
   because the context will not actually be used and is expensive
   to obtain.
 - When creating values, don't use v8::TryCatch (NAPI_PREAMBLE),
   because these functions have no possibility of executing JS code.

Refs: https://github.com/nodejs/node/issues/14379
PR-URL: https://github.com/nodejs/node/pull/14573

+++ test/addons-napi/7_factory_wrap/myobject.h
@@ -18,7 +18,7 @@ class MyObject {
   static napi_ref constructor;
   static napi_value New(napi_env env, napi_callback_info info);
   static napi_value PlusOne(napi_env env, napi_callback_info info);
-  double counter_;
+  uint32_t counter_;
   napi_env env_;
   napi_ref wrapper_;
 };
